backend/requirements.txt
flask
flask-cors
transformers
torch
google-generativeai
python-dotenv
